PSR5 prod (#280)
* merge main into develop (#266)

* fi Fix react-error-overlay dependency (#267)

* Music, "Apply to" button, legal pages (#270)

* header - "apply to" button

* bg music

* legal - t&c

* privacy page

* format

* nit

* nit (#273)

* PSR 3 (#274)

* page loader - images only

* add videos, fonts and timeout to PageLoader

* improve page loader

* page size limit

* fix page size limit

* better PageSizeLimit for mobile

* coming-soon mobile

* remove local video

* format

* PSR 4 (#276)

* fix "landscape" when keyboard is open in mobile

* fix coming soon css

* add new color

* format

* Fix diamond object to use uint64 for GIA number (#278)

* about - wip (#279)

+++ smart-contracts/contracts/interface/IDiamondDawnMine.sol
@@ -9,7 +9,7 @@ interface IDiamondDawnMine {
     event Mine(uint tokenId);
     event Cut(uint tokenId);
     event Polish(uint tokenId);
-    event Ship(uint tokenId, uint16 physicalId, uint32 number);
+    event Ship(uint tokenId, uint16 physicalId, uint64 number);
     event Dawn(uint tokenId);
 
     function initialize(uint16 maxDiamond) external;
